refactor: apis/league 폴더구조 변경
