+++ src/components/common/account/selectbox.tsx
@@ -5,7 +5,7 @@ import { useEffect, useRef, useState } from 'react';
 import OptionItem, { Option } from '@/components/common/option-item';
 import clsx from 'clsx';
 import { TeamDto } from '@/services/apis/team/team.type';
-import { LeagueDto } from '@/services/apis/league/dto';
+import { LeagueDto } from '@/services/apis/league/league.type';
 import { NO_CHEERING_TEAM_PK } from '@/lib/constants/noCheeringTeam';
 import { useCurrentUserInfoStore } from '@/lib/store/useCurrentUserInfoStore';
 import AlertModal from '@/components/features/detail/alert-modal';
