+++ src/components/layouts/with-side/ranking-list/ranking-list.tsx
@@ -5,7 +5,7 @@ import RankingItem from './ranking-item';
 import SelectBox from './select-box';
 import { getActualSeasonRanking, getGambleSeasonRanking } from '@/services/apis/ranking/ranking.api';
 import FetchingFailedCard from '@/components/common/fetching-failed-card';
-import { LeagueDto } from '@/services/apis/league/dto';
+import { LeagueDto } from '@/services/apis/league/league.type';
 import { useCallback, useEffect, useState } from 'react';
 import { ActualRankingDto, GambleRankingDto } from '@/services/apis/ranking/ranking.type';
 import { useCurrentUserInfoStore } from '@/lib/store/useCurrentUserInfoStore';
