+++ src/components/layouts/with-side/ranking-list/select-box.tsx
@@ -4,8 +4,8 @@ import Image from 'next/image';
 import { useEffect, useRef, useState } from 'react';
 import OptionItem from '../../../common/option-item';
 import clsx from 'clsx';
-import { LeagueDto } from '@/services/apis/league/dto';
-import { getLeague } from '@/services/apis/league';
+import { LeagueDto } from '@/services/apis/league/league.type';
+import { getLeague } from '@/services/apis/league/league.api';
 
 export default function SelectBox({
 	content,
